GPS_TrackingSystem_Backend/requirements.txt
Django==4.2.28
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.0
PyJWT==2.12.1
requests==2.31.0
